combine: writeNeo4j and readNeo4j in Neo4jService
