+++ src/main/kotlin/ioNeo4j/RepoNeo4j.kt
@@ -1,0 +1,7 @@
+package ioNeo4j
+
+import org.springframework.data.neo4j.repository.Neo4jRepository
+
+interface ReadRepositoryNeo4j : Neo4jRepository<VertexNeo4j, Long>
+
+interface WriteRepositoryNeo4j : Neo4jRepository<VertexNeo4j, Long>
